move BootInfo to the hardware dependent part
