+++ src/arch/x86_64/mod.rs
@@ -1,13 +1,87 @@
 pub mod registers;
 
 use core::arch::x86_64::_rdtsc as rdtsc;
+use goblin::elf64::header::EM_X86_64;
 use log::{debug, warn};
 use raw_cpuid::CpuId;
 use std::convert::TryInto;
 use std::time::{Duration, Instant};
 
+pub const ELF_HOST_ARCH: u16 = EM_X86_64;
 const MHZ_TO_HZ: u64 = 1000000;
 const KHZ_TO_HZ: u64 = 1000;
+
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+pub struct BootInfo {
+	pub magic_number: u32,
+	pub version: u32,
+	pub base: u64,
+	pub limit: u64,
+	pub image_size: u64,
+	pub tls_start: u64,
+	pub tls_filesz: u64,
+	pub tls_memsz: u64,
+	pub current_stack_address: u64,
+	pub current_percore_address: u64,
+	pub host_logical_addr: u64,
+	pub boot_gtod: u64,
+	pub mb_info: u64,
+	pub cmdline: u64,
+	pub cmdsize: u64,
+	pub cpu_freq: u32,
+	pub boot_processor: u32,
+	pub cpu_online: u32,
+	pub possible_cpus: u32,
+	pub current_boot_id: u32,
+	pub uartport: u16,
+	pub single_kernel: u8,
+	pub uhyve: u8,
+	pub hcip: [u8; 4],
+	pub hcgateway: [u8; 4],
+	pub hcmask: [u8; 4],
+	pub tls_align: u64,
+}
+
+impl BootInfo {
+	pub fn new() -> Self {
+		BootInfo {
+			magic_number: 0xC0DE_CAFEu32,
+			version: 1,
+			base: 0,
+			limit: 0,
+			tls_start: 0,
+			tls_filesz: 0,
+			tls_memsz: 0,
+			image_size: 0,
+			current_stack_address: 0,
+			current_percore_address: 0,
+			host_logical_addr: 0,
+			boot_gtod: 0,
+			mb_info: 0,
+			cmdline: 0,
+			cmdsize: 0,
+			cpu_freq: 0,
+			boot_processor: !0,
+			cpu_online: 0,
+			possible_cpus: 0,
+			current_boot_id: 0,
+			uartport: 0,
+			single_kernel: 1,
+			uhyve: 0,
+			hcip: [255, 255, 255, 255],
+			hcgateway: [255, 255, 255, 255],
+			hcmask: [255, 255, 255, 0],
+			tls_align: 0,
+		}
+	}
+}
+
+impl Default for BootInfo {
+	fn default() -> Self {
+		Self::new()
+	}
+}
 
 pub fn detect_freq_from_cpuid(cpuid: &CpuId) -> std::result::Result<u32, ()> {
 	debug!("Trying to detect CPU frequency by tsc info");
