+++ src/macos/aarch64/uhyve.rs
@@ -1,6 +1,7 @@
+use crate::aarch64::BootInfo;
 use crate::macos::aarch64::vcpu::*;
 use crate::vm::HypervisorResult;
-use crate::vm::{BootInfo, Parameter, Vm};
+use crate::vm::{Parameter, Vm};
 use libc;
 use libc::c_void;
 use log::debug;
